+++ src/main/kotlin/store/pengu/server/data/ShoppingList.kt
@@ -3,6 +3,6 @@ package store.pengu.server.data
 data class ShoppingList(
     override val id: Long,
     override val name: String,
-    override val latitude: Float,
-    override val longitude: Float,
+    override val latitude: Double,
+    override val longitude: Double,
 ) : UserList(id, name, latitude, longitude)
