+++ src/main/kotlin/store/pengu/server/data/UserList.kt
@@ -3,6 +3,6 @@ package store.pengu.server.data
 abstract class UserList(
     open val id: Long,
     open val name: String,
-    open val latitude: Float,
-    open val longitude: Float
+    open val latitude: Double,
+    open val longitude: Double
 )
