+++ src/main/resources/db/migrations/V11__create_local_product_prices_table.sql
@@ -2,7 +2,7 @@ create table local_product_prices (
     product_id BIGINT UNSIGNED,
         FOREIGN KEY (product_id) REFERENCES products(id) ON UPDATE CASCADE ON DELETE CASCADE ,
     price DOUBLE NOT NULL,
-    latitude FLOAT NOT NULL,
-    longitude FLOAT NOT NULL,
+    latitude DOUBLE NOT NULL,
+    longitude DOUBLE NOT NULL,
     PRIMARY KEY (product_id, latitude, longitude)
 );
